back-end/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
markdown2 @ file:///home/conda/feedstock_root/build_artifacts/markdown2_1638543728645/work
Pillow==9.1.0
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
PyJWT==2.4.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
tzdata==2022.1
urllib3==1.26.9
wincertstore==0.2
